requirements.txt
pyjwt
tqdm
termcolor
requests
